+++ client/helpers/handlebars.js
@@ -15,7 +15,12 @@ UI.registerHelper('eachWithRank', function(items, options) {
 });
 
 UI.registerHelper('getSetting', function(setting, defaultArgument){
-  return getSetting(setting, defaultArgument);
+  var defaultArgument = (typeof defaultArgument !== 'undefined') ? defaultArgument : '';
+  var setting = getSetting(setting, defaultArgument);
+  return setting;
+});
+UI.registerHelper('isLoggedIn', function() {
+  return !!Meteor.user();
 });
 UI.registerHelper('canView', function() {
   return canView(Meteor.user());
@@ -52,4 +57,8 @@ UI.registerHelper('canEdit', function(collectionName, item, action) {
 
 UI.registerHelper('log', function(context){
   console.log(context);
+});
+
+UI.registerHelper("formatDate", function(datetime, format) {
+  return moment(datetime).format(format);
 });