+++ client/views/common/css.js
@@ -1,0 +1,2 @@
+Template[getTemplate('css')].helpers({
+});