+++ client/views/common/footer.js
@@ -4,5 +4,11 @@ Template[getTemplate('footer')].helpers({
   }, 
   footerClass: function(){
     return Session.get('isPostsList') ? 'absolute' : 'static';
+  },
+  footerModules: function () {
+    return footerModules;
+  },
+  getTemplate: function () {
+    return getTemplate(this.template);
   }
 });