+++ client/views/users/user_email.js
@@ -1,6 +1,9 @@
 Template[getTemplate('user_email')].helpers({
   user: function(){
     return Meteor.user();
+  },
+  username: function () {
+    return getUserName(Meteor.user());
   }
 });
 
@@ -26,7 +29,7 @@ Template[getTemplate('user_email')].events({
         throwError(error.reason);
       } else {
         throwError(i18n.t('Thanks for signing up!'));
-        Meteor.call('addCurrentUserToMailChimpList');
+        // Meteor.call('addCurrentUserToMailChimpList');
         trackEvent("new sign-up", {'userId': user._id, 'auth':'twitter'});
         Router.go('/');
       }