+++ collections/notifications.js
@@ -29,70 +29,60 @@ Notifications.allow({
   , remove: canEditById
 });
 
-getNotificationContents = function(notification, context){
-  // the same notifications can be displayed in multiple contexts: on-site in the sidebar, sent by email, etc. 
+createNotification = function(event, properties, userToNotify) {
+  // 1. Store notification in database
+  var notification = {
+    timestamp: new Date().getTime(),
+    userId: userToNotify._id,
+    event: event,
+    properties: properties,
+    read: false
+  };
+  var newNotificationId=Notifications.insert(notification);
+
+  // 2. Send notification by email (if on server)
+  if(Meteor.isServer && getUserSetting('notifications.replies', false, userToNotify)){
+    // put in setTimeout so it doesn't hold up the rest of the method
+    Meteor.setTimeout(function () {
+      notificationEmail = buildEmailNotification(notification);
+      sendEmail(getEmail(userToNotify), notificationEmail.subject, notificationEmail.html);
+    }, 1);
+  }
+};
+
+buildSiteNotification = function (notification) {
   var event = notification.event,
-      p = notification.properties,
-      context = typeof context === 'undefined' ? 'sidebar' : context,
-      userToNotify = Meteor.users.findOne(notification.userId);
+      comment = notification.properties.comment,
+      post = notification.properties.post,
+      userToNotify = Meteor.users.findOne(notification.userId),
+      template,
+      html
+
+  var properties = {
+    profileUrl: getProfileUrlById(comment.userId),
+    author: comment.author,
+    postCommentUrl: getPostCommentUrl(post._id, comment._id),
+    postTitle: post.title
+  }
 
   switch(event){
     case 'newReply':
-      var n = {
-        subject: 'Someone replied to your comment on "'+p.postTitle+'"',
-        text: p.commentAuthorName+' has replied to your comment on "'+p.postTitle+'": '+getPostCommentUrl(p.postId, p.commentId),
-        html: '<p><a href="'+getProfileUrlById(p.commentAuthorId)+'">'+p.commentAuthorName+'</a> has replied to your comment on "<a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+p.postTitle+'</a>"</p>'
-      };
-      if(context == 'email')
-        n.html += '<p>'+p.commentExcerpt+'</p><a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">Read more</a>';
+      template = 'notification_new_reply';
       break;
 
     case 'newComment':
-      var n = {
-        subject: 'A new comment on your post "'+p.postTitle+'"',
-        text: 'You have a new comment by '+p.commentAuthorName+' on your post "'+p.postTitle+'": '+getPostCommentUrl(p.postId, p.commentId),
-        html: '<p><a href="'+getProfileUrlById(p.commentAuthorId)+'">'+p.commentAuthorName+'</a> left a new comment on your post "<a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+p.postTitle+'</a>"</p>'
-      };
-      if(context == 'email')
-        n.html += '<p>'+p.commentExcerpt+'</p><a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">Read more</a>';
-      break;
-
-    case 'newPost':
-      var n = {
-        subject: p.postAuthorName+' has created a new post: "'+p.postTitle+'"',
-        text: p.postAuthorName+' has created a new post: "'+p.postTitle+'" '+getPostUrl(p.postId),
-        html: '<a href="'+getProfileUrlById(p.postAuthorId)+'">'+p.postAuthorName+'</a> has created a new post: "<a href="'+getPostUrl(p.postId)+'" class="action-link">'+p.postTitle+'</a>".'  
-      };
-      break;
-
-    case 'accountApproved':
-      var n = {
-        subject: 'Your account has been approved.',
-        text: 'Welcome to '+getSetting('title')+'! Your account has just been approved.',
-        html: 'Welcome to '+getSetting('title')+'!<br/> Your account has just been approved. <a href="'+Meteor.absoluteUrl()+'">Start posting.</a>'
-      };
-      break;
-
-    case 'newUser':
-      var n = {
-        subject: 'New user: '+p.username,
-        text: 'A new user account has been created: '+p.profileUrl,
-        html: 'A new user account has been created: <a href="'+p.profileUrl+'">'+p.username+'</a>'
-      };
-      break;
+      template = 'notification_new_comment';
+      break; 
 
     default:
-    break;
+      break;
   }
 
-  // if context is email, append unsubscribe link to all outgoing notifications
-  if(context == 'email'){
-    n.to = getEmail(userToNotify);
-    n.text = n.text + '\n\n Unsubscribe from all notifications: '+getUnsubscribeLink(userToNotify);
-    n.html = n.html + '<br/><br/><a href="'+getUnsubscribeLink(userToNotify)+'">Unsubscribe from all notifications</a>';
-  }
+  html = Blaze.toHTML(Blaze.With(properties, function(){
+    return Template[getTemplate(template)]
+  }));
 
-  return n;
+  return html;
 };
 
 Meteor.methods({