+++ packages/telescope-base/lib/base.js
@@ -31,8 +31,9 @@ adminNav = [];
 // array containing subscriptions to be preloaded
 preloadSubscriptions = [];
 
-// array containing nav items; initialize with views menu
-navItems = ['viewsMenu'];
+// array containing nav items; initialize with views menu and admin menu
+navItems = ['viewsMenu', 'adminMenu'];
+
 
 // object containing post list view parameters
 viewParameters = {}
@@ -75,6 +76,10 @@ viewParameters.digest = function (terms) {
     }
   };
 }
+
+footerModules = [];
+
+heroModules = [];
 
 // array containing post modules
 modulePositions = [
@@ -135,9 +140,11 @@ postMeta = [
 ]
 // ------------------------------ Callbacks ------------------------------ //
 
+postSubmitRenderedCallbacks = [];
 postSubmitClientCallbacks = [];
 postSubmitServerCallbacks = [];
 
+postEditRenderedCallbacks = [];
 postEditClientCallbacks = [];
 
 commentEditClientCallbacks = []; // not used yet