+++ packages/telescope-lib/package.js
@@ -19,6 +19,7 @@ Package.on_use(function (api) {
     'camelToDash',
     'dashToCamel',
     'getSetting',
-    'getSiteUrl'
+    'getSiteUrl',
+    'trimWords'
   ]);
 });