+++ packages/telescope-rss/lib/server/rss.js
@@ -10,10 +10,10 @@ serveRSS = function() {
   Posts.find({status: STATUS_APPROVED}, {sort: {postedAt: -1}, limit: 20}).forEach(function(post) {
     feed.item({
      title: post.title,
-     description: post.body+'</br></br> <a href="'+getPostUrl(post._id)+'">Comments</a>',
+     description: post.body+'</br></br> <a href="'+getPostPageUrl(post._id)+'">Comments</a>',
      author: post.author,
      date: post.postedAt,
-     url: (post.url ? getOutgoingUrl(post.url) : getPostUrl(post._id)),
+     url: getPostLink(post),
      guid: post._id
     });
   });