+++ packages/telescope-theme-hubble/package.js
@@ -1,17 +1,9 @@
-Package.describe("Telescope Hubble theme");
+Package.describe({summary: "Telescope Hubble theme"});
 
 Package.on_use(function (api) {
 
-  // api.use(['telescope-lib'], ['client', 'server']);
-
-  // api.use([
-  //   'jquery',
-  //   'underscore',
-  //   'templating'
-  // ], 'client');
-
   api.add_files([
-    'lib/client/stylesheets/screen.css',
+    'lib/client/css/screen.css',
     ], ['client']);
   
 });