+++ server/users.js
@@ -40,8 +40,12 @@ Accounts.onCreateUser(function(options, user){
   trackEvent('new user', {username: user.username, email: user.profile.email});
 
   // if user has already filled in their email, add them to MailChimp list
-  if(user.profile.email)
-    addToMailChimpList(user);
+  // if(user.profile.email)
+  //   addToMailChimpList(user, false, function(error, result){
+  //     if(error){
+  //       console.log(error)
+  //     }
+  //   });
 
   // if the new user has been invited 
   // set her status accordingly and update invitation info
@@ -66,15 +70,12 @@ Accounts.onCreateUser(function(options, user){
   var admins = Meteor.users.find({isAdmin: true});
   admins.forEach(function(admin){
     if(getUserSetting('notifications.users', false, admin)){
-      var notification = getNotificationContents({
-        event: 'newUser',
-        properties: {
-          username: getUserName(user),
-          profileUrl: getProfileUrl(user)
-        },
-        userId: admin._id
-      }, 'email');
-      sendNotification(notification, admin);
+      var emailProperties = {
+        profileUrl: getProfileUrl(user),
+        username: getUserName(user)
+      }
+      var html = Handlebars.templates[getTemplate('emailNewUser')](emailProperties);
+      sendEmail(getEmail(admin), 'New user account: '+getUserName(user), buildEmailTemplate(html));
     }
   });
 
@@ -84,27 +85,6 @@ Accounts.onCreateUser(function(options, user){
 getEmailHash = function(user){
   // todo: add some kind of salt in here
   return CryptoJS.MD5(getEmail(user).trim().toLowerCase() + user.createdAt).toString();
-};
-
-addToMailChimpList = function(user){
-  // add a user to a MailChimp list.
-  // called when a new user is created, or when an existing user fills in their email
-  if((MAILCHIMP_API_KEY=getSetting('mailChimpAPIKey')) && (MAILCHIMP_LIST_ID=getSetting('mailChimpListId'))){
-
-    var email = getEmail(user);
-    if (! email)
-      throw 'User must have an email address';
-
-    console.log('adding "'+email+'" to MailChimp list…');
-    
-    var mailChimp = new MailChimpAPI(MAILCHIMP_API_KEY, { version : '1.3', secure : false });
-    
-    mailChimp.listSubscribe({
-      id: MAILCHIMP_LIST_ID,
-      email_address: email,
-      double_optin: false
-    });
-  }
 };
 
 Meteor.methods({
@@ -133,8 +113,5 @@ Meteor.methods({
   setEmailHash: function(user){
     var email_hash = CryptoJS.MD5(getEmail(user).trim().toLowerCase()).toString();
     Meteor.users.update(user._id, {$set : {email_hash : email_hash}});
-  },
-  addCurrentUserToMailChimpList: function(){
-    addToMailChimpList(Meteor.user());
   }
 });